backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-environ
psycopg2-binary
pandas
setuptools
coreapi
Pillow
pyrebase4
django-environ
django-storages 
google-cloud-storage
firebase-admin
django-resumable
